Tidy up #5
Splitting glove_measurements.dart into multiple files.
