+++ lib/datacollection/storage.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 import 'dart:convert';
+import 'package:lsa_gloves/model/finger.dart';
 import 'package:lsa_gloves/model/glove_measurement.dart';
 import 'package:lsa_gloves/edgeimpulse/api_client.dart';
 import 'package:path_provider/path_provider.dart';
