+++ lib/model/glove_measurement.dart
@@ -1,29 +1,6 @@
+import 'finger.dart';
 
-enum FingerValue {
-  Pinky,
-  Ring,
-  Middle,
-  Index,
-  Thumb
-}
-
-extension FingerValueTranslation on FingerValue {
-  String spanishName() {
-    switch (this) {
-      case FingerValue.Pinky:
-        return "Meñique";
-      case FingerValue.Ring:
-        return "Anular";
-      case FingerValue.Middle:
-        return "Medio";
-      case FingerValue.Index:
-        return "Índice";
-      case FingerValue.Thumb:
-        return "Pulgar";
-    }
-  }
-}
-
+/// Class to represent a glove measurement.
 class GloveMeasurement {
   static const int measurementsNumber = 6;
   static const String pinkyLetter = "P";
@@ -99,129 +76,4 @@ class GloveMeasurement {
   }
 }
 
-enum SensorValue {
-  Acceleration,
-  Gyroscope
-}
-
-extension SensorValueExtension on SensorValue {
-  String spanishName() {
-    switch (this) {
-      case SensorValue.Acceleration:
-        return "Acelerómetro";
-      case SensorValue.Gyroscope:
-        return "Giroscopio";
-    }
-  }
-
-  String getXLabel() {
-    switch (this) {
-      case SensorValue.Acceleration:
-        return "x (m/s²)";
-      case SensorValue.Gyroscope:
-        return "x (º/s)";
-    }
-  }
-  String getYLabel() {
-    switch (this) {
-      case SensorValue.Acceleration:
-        return "y (m/s²)";
-      case SensorValue.Gyroscope:
-        return "y (º/s)";
-    }
-  }
-  String getZLabel() {
-    switch (this) {
-      case SensorValue.Acceleration:
-        return "z (m/s²)";
-      case SensorValue.Gyroscope:
-        return "z (º/s)";
-    }
-  }
-}
-
-class Finger {
-  final Acceleration acc;
-  final Gyro gyro;
-
-  Finger(this.acc, this.gyro);
-
-  Finger.fromJson(Map<String, dynamic> json)
-      : acc = Acceleration.fromJson(json['acc'] as Map<String, dynamic>),
-        gyro = Gyro.fromJson(json['gyro'] as Map<String, dynamic>);
-      Map<String, dynamic> toJson() => {
-    'acc': acc.toJson(),
-    'gyro': gyro.toJson(),
-  };
-
-  Finger.fromList(List<double> m):
-        acc = Acceleration(m[0],m[1], m[2]),
-        gyro = Gyro(m[3],m[4], m[5]);
-
-  Vector3 getSensorValues(SensorValue sensorName) {
-    switch (sensorName) {
-      case SensorValue.Acceleration:
-        return acc;
-      case SensorValue.Gyroscope:
-        return gyro;
-    }
-  }
-}
-
-class Acceleration with Vector3 {
-  final double x;
-  final double y;
-  final double z;
-
-  Acceleration(this.x, this.y, this.z);
-
-  Acceleration.fromJson(Map<String, dynamic> json)
-      : x = json['x'], y = json['y'], z = json['z'];
-
-  Map<String, dynamic> toJson() => {
-    'x': x,
-    'y': y,
-    'z': z,
-  };
-
-  @override
-  double getX() => x;
-
-  @override
-  double getY() => y;
-
-  @override
-  double getZ() => z;
-}
-
-class Gyro with Vector3 {
-  final double x;
-  final double y;
-  final double z;
-  Gyro (this.x, this.y, this.z);
-
-  Gyro .fromJson(Map<String, dynamic> json)
-      : x = json['x'], y = json['y'], z = json['z'];
-
-  Map<String, dynamic> toJson() => {
-    'x': x,
-    'y': y,
-    'z': z,
-  };
-
-  @override
-  double getX() => x;
-
-  @override
-  double getY() => y;
-
-  @override
-  double getZ() => z;
-}
-
-abstract class Vector3 {
-  double getX();
-  double getY();
-  double getZ();
-}
