+++ lib/model/vector3.dart
@@ -1,0 +1,6 @@
+/// Helper class to contain the 3 axis values of the mpu6050 sensors.
+abstract class Vector3 {
+  double getX();
+  double getY();
+  double getZ();
+}
