+++ lib/pages/data_visualization_page.dart
@@ -6,7 +6,10 @@ import 'package:intl/intl.dart';
 import 'package:lsa_gloves/connection/ble/bluetooth_backend.dart';
 import 'package:lsa_gloves/datacollection/measurements_collector.dart';
 import 'package:lsa_gloves/datacollection/measurements_listener.dart';
+import 'package:lsa_gloves/model/finger.dart';
 import 'package:lsa_gloves/model/glove_measurement.dart';
+import 'package:lsa_gloves/model/sensor_value.dart';
+import 'package:lsa_gloves/model/vector3.dart';
 import 'package:lsa_gloves/navigation/navigation_drawer.dart';
 import 'package:provider/provider.dart';
 import 'package:syncfusion_flutter_charts/charts.dart';
