+++ lib/pages/file_content_chart_page.dart
@@ -2,7 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:lsa_gloves/connection/ble/bluetooth_backend.dart';
 import 'package:lsa_gloves/datacollection/storage.dart';
+import 'package:lsa_gloves/model/acceleration.dart';
+import 'package:lsa_gloves/model/finger.dart';
 import 'package:lsa_gloves/model/glove_measurement.dart';
+import 'package:lsa_gloves/model/gyro.dart';
+import 'package:lsa_gloves/model/sensor_value.dart';
+import 'package:lsa_gloves/model/vector3.dart';
 import 'package:lsa_gloves/navigation/navigation_drawer.dart';
 import 'package:provider/provider.dart';
 import 'package:syncfusion_flutter_charts/charts.dart';
